ai_platform_engineering/utils/pyproject.toml
[project]
name = "ai-platform-engineering-utils"
version = "0.2.0"
license = "Apache-2.0"
description = "Common utilities and base classes for AI Platform Engineering agents"
readme = "README.md"
authors = [
    {name = "CNOE Team", email = "info@cnoe.io"},
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "a2a-sdk==0.3.0",
    "agntcy-app-sdk==0.4.5",
    "cnoe-agent-utils==0.3.14",
    "httpx==0.28.1",
    "langchain-core==1.2.31",
    "langchain-mcp-adapters==0.2.1",
    "langgraph==1.1.6",
    "langgraph-checkpoint-mongodb==0.3.1",
    "langmem==0.0.30",
    "mcp==1.26.0",
    "prometheus-client==0.23.1",
    "pydantic==2.12.5",
    "pyjwt==2.12.0",
    "pymongo==4.15.5",
    "python-dotenv==1.2.1",
    "beautifulsoup4==4.14.3",
    "requests==2.33.0",
    "strands-agents==1.19.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 140
indent-width = 2

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
]
ignore = ["F403"]

[tool.uv]
override-dependencies = ["protobuf==5.29.6"]
constraint-dependencies = ["nltk==3.9.4", "pypdf==6.10.2"]
